Split into read and write halves
